requirements.txt
google-genai
python-dotenv
ddgs
docker
requests
faster-whisper
piper-tts
pillow
croniter